+++ test/unit/rx-collection.test.ts
@@ -138,11 +138,11 @@ describe('rx-collection.test.ts', () => {
                     const lastIndexDefFields = indexes.indexes[1].def.fields;
                     assert.deepStrictEqual(
                         lastIndexDefFields, [
-                            { 'age': 'asc' },
-                            { '|a': 'asc' },
-                            // the primaryKey index will always be added by RxDB
-                            { _id: 'asc' }
-                        ]
+                        { 'age': 'asc' },
+                        { '|a': 'asc' },
+                        // the primaryKey index will always be added by RxDB
+                        { _id: 'asc' }
+                    ]
                     );
                     db.destroy();
                 });
@@ -1102,7 +1102,8 @@ describe('rx-collection.test.ts', () => {
                          * because there we use the in-memory persistence
                          * which is not really persistent between different writes.
                          */
-                        config.storage.name === 'dexie-worker'
+                        config.storage.name === 'dexie-worker' ||
+                        config.storage.name === 'remote'
                     ) {
                         return;
                     }
@@ -1426,8 +1427,12 @@ describe('rx-collection.test.ts', () => {
         });
         config.parallel('.bulkUpsert()', () => {
             it('insert and update', async () => {
+                console.log('---- 0');
+
                 const c = await humansCollection.create(0);
                 const amount = 5;
+
+                console.log('---- 1');
 
                 // insert
                 await c.bulkUpsert(
@@ -1435,6 +1440,7 @@ describe('rx-collection.test.ts', () => {
                 );
                 let allDocs = await c.find().exec();
                 assert.strictEqual(allDocs.length, 5);
+                console.log('---- 2');
 
                 // update
                 const docsData = allDocs.map(d => {
@@ -1442,10 +1448,12 @@ describe('rx-collection.test.ts', () => {
                     data.age = 100;
                     return data;
                 });
+                console.log('---- 3');
                 await c.bulkUpsert(docsData);
                 allDocs = await c.find().exec();
                 assert.strictEqual(allDocs.length, 5);
                 allDocs.forEach(d => assert.strictEqual(d.age, 100));
+                console.log('---- 4');
 
                 c.database.destroy();
             });